requirements.txt
flask
torch
torchvision
requests
Pillow
beautifulsoup4
matplotlib
gunicorn==21.2.0
